+++ CHANGES.md
@@ -10,6 +10,7 @@
 * Moved `_dst_geom_in_tms_crs` from Reader to `SpatialMixin` class **breaking change**
 * Removed use of rasterio's `is_tiled` method
 * Enable **Alternate** asset's HREF for STAC by using `RIO_TILER_STAC_ALTERNATE_KEY` environment variable
+* Adding support for GDAL VRT Connection string for STAC Assets
 
 # 6.7.0 (2024-09-05)